[AN] refactor: CategoryComposeFragment 패키지 이동 및 데이터 갱신 로직 수정 #545
