+++ android/app/src/main/java/com/onair/hearit/presentation/search/SearchFragment.kt
@@ -14,7 +14,10 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.onair.hearit.R
 import com.onair.hearit.databinding.FragmentSearchBinding
-import com.onair.hearit.presentation.CategoryComposeFragment
+import com.onair.hearit.presentation.IntentKeys.CATEGORY_COLOR_KEY
+import com.onair.hearit.presentation.IntentKeys.CATEGORY_ID_KEY
+import com.onair.hearit.presentation.IntentKeys.CATEGORY_NAME_KEY
+import com.onair.hearit.presentation.search.category.CategoryComposeFragment
 import com.onair.hearit.presentation.search.recent.SearchRecentFragment
 
 class SearchFragment :
@@ -102,9 +105,9 @@ class SearchFragment :
             CategoryComposeFragment().apply {
                 arguments =
                     bundleOf(
-                        "CATEGORY_ID" to id,
-                        "CATEGORY_NAME" to name,
-                        "CATEGORY_COLOR" to colorCode,
+                        CATEGORY_ID_KEY to id,
+                        CATEGORY_NAME_KEY to name,
+                        CATEGORY_COLOR_KEY to colorCode,
                     )
             }
